pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "water-network-anomaly-detection"
version = "1.0.0"
description = "Water Network Anomaly Detection using Hybrid ML (KNN + LSTM)"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Water Analytics Team", email = "team@example.com"}
]
keywords = [
    "water-network",
    "anomaly-detection",
    "machine-learning",
    "knn",
    "lstm",
    "synthetic-data"
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "tensorflow>=2.10.0",
    "joblib>=1.1.0",
    "python-dotenv>=0.19.0",
    "chromadb>=0.3.0",
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langgraph>=0.1.0",
    "langchain-text-splitters>=0.0.1",
    "tiktoken>=0.5.0",
    "pypdf>=3.0.0",
    "python-docx>=0.8.11",
    "httpx>=0.24.0",
    "urllib3>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "isort>=5.0.0",
]

rag-llm-providers = [
    "langchain-openai>=0.0.1",
    "langchain-anthropic>=0.0.1",
    "langchain-huggingface>=0.0.1",
    "langchain-ollama>=0.0.1",
]

[project.urls]
"Repository" = "https://github.com/sopradhan/water"
"Bug Tracker" = "https://github.com/sopradhan/water/issues"

[tool.setuptools]
packages = ["src", "src.config", "src.data_generator", "src.model", "src.rag", "src.rag.agents", "src.rag.agents.langgraph_agent", "src.rag.agents.healing_agent", "src.rag.config", "src.rag.guardrails", "src.rag.reporting", "src.rag.tools", "src.rag.tools.services", "src.rag.pages", "src.rag.api"]

[tool.setuptools.package-data]
src = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--strict-markers"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
]

[tool.coverage.run]
branch = true
source = ["src"]
omit = [
    "*/tests/*",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
